Fix admin credentials auth configuration
