+++ nerin-electric-site-v3-fixed/middleware.ts
@@ -6,5 +6,5 @@ export const middleware: NextMiddleware = (request: NextRequest) => baseMiddlewa
 export default middleware
 
 export const config = {
-  matcher: ['/((?!api/health|_next/static|_next/image|favicon.ico|robots.txt|sitemap.xml).*)'],
+  matcher: ['/((?!api/auth|api/health|_next/static|_next/image|favicon.ico|robots.txt|sitemap.xml).*)'],
 }
